Move AuthMiddleware to authorizations app
This fixes applications not including the `vng_api_common.authorizations` app to create duplicate migrations. See https://github.com/maykinmedia/open-api-framework/pull/70#discussion_r1768287334
